Remove old Product overview codes and move still in used files to the new folder.
