+++ client/a8c-for-agencies/sections/marketplace/products-overview-v2/hooks/use-assign-licenses-to-site.ts
