+++ client/a8c-for-agencies/sections/marketplace/products-overview-v2/index.tsx
@@ -26,10 +26,10 @@ import { PRODUCT_FILTER_KEY_CATEGORIES } from '../constants';
 import { MarketplaceTypeContext, ShoppingCartContext } from '../context';
 import withMarketplaceType from '../hoc/with-marketplace-type';
 import useShoppingCart from '../hooks/use-shopping-cart';
-import { useProductBundleSize } from '../products-overview/product-listing/hooks/use-product-bundle-size';
-import useSelectedProductFilters from '../products-overview/product-listing/hooks/use-selected-product-filters';
 import ShoppingCart from '../shopping-cart';
 import useCompactOnScroll from './hooks/use-compact-on-scroll';
+import { useProductBundleSize } from './hooks/use-product-bundle-size';
+import useSelectedProductFilters from './hooks/use-selected-product-filters';
 import ProductActionPanel from './product-action-panel';
 import ProductCategoryMenu from './product-category-menu';
 import ProductListing from './product-listing';
