+++ client/a8c-for-agencies/sections/marketplace/products-overview-v2/product-card/index.tsx
@@ -207,6 +207,7 @@ function ProductCard( props: Props ) {
 							customText={ translate( 'View details' ) }
 							productName={ getProductShortTitle( currentProduct ) }
 							onClick={ onShowLightbox }
+							showIcon={ false }
 						/>
 					) }
 				</div>
