+++ client/a8c-for-agencies/sections/marketplace/products-overview-v2/product-listing/index.tsx
@@ -14,8 +14,8 @@ import { recordTracksEvent } from 'calypso/state/analytics/actions';
 import { ShoppingCartContext } from '../../context';
 import useProductAndPlans from '../../hooks/use-product-and-plans';
 import { SelectedFilters } from '../../lib/product-filter';
-import { getSupportedBundleSizes } from '../../products-overview/product-listing/hooks/use-product-bundle-size';
-import useSubmitForm from '../../products-overview/product-listing/hooks/use-submit-form';
+import { getSupportedBundleSizes } from '../hooks/use-product-bundle-size';
+import useSubmitForm from '../hooks/use-submit-form';
 import ProductCard from '../product-card';
 import ProductListingEmpty from './empty';
 import ProductListingSection from './section';
