--- client/a8c-for-agencies/sections/marketplace/products-overview/product-listing/style.scss
@@ -1,117 +1,0 @@
-@import "@wordpress/base-styles/breakpoints";
-@import "@wordpress/base-styles/mixins";
-@import "../../mixins";
-
-.product-listing__placeholder {
-	@include placeholder( --color-neutral-10 );
-	height: 43px;
-}
-
-p.product-listing__description {
-	flex: 1 1 auto;
-	align-self: flex-end;
-	margin: 1rem 0;
-	font-size: 0.875rem;
-	color: #333;
-
-	@include breakpoint-deprecated( ">660px" ) {
-		margin-right: 1rem;
-	}
-}
-
-.product-listing__select-license {
-	// :active introduces a 1px border but :focus introduces a 1.5px border which causes the button to move around.
-	border-width: 1.5px !important;
-	border-color: transparent;
-}
-
-.product-listing__actions {
-	display: flex;
-	flex-wrap: wrap;
-	gap: 16px;
-	margin: 16px 0 32px;
-	justify-content: space-between;
-	position: relative;
-	z-index: 1;
-}
-
-.select-dropdown.is-compact.product-listing__product-filter-select {
-	flex-basis: 100%;
-	height: 46px;
-
-	@include break-xlarge {
-		flex-basis: auto;
-	}
-
-	.select-dropdown__header {
-		height: 46px;
-	}
-
-	@include breakpoint-deprecated( ">660px" ) {
-		height: 35px;
-
-		.select-dropdown__header {
-			height: 35px;
-		}
-	}
-
-	.select-dropdown__header-text {
-		font-size: rem(13px);
-		font-weight: 400;
-		color: var(--color-text);
-		padding-inline-end: 8px;
-	}
-
-	.select-dropdown__header-text b {
-		color: var(--color-text);
-	}
-
-	.select-dropdown__container {
-		width: 100%;
-
-		@include break-xlarge {
-			width: auto;
-		}
-	}
-
-	.select-dropdown__item.is-selected {
-		background: var(--color-link-5);
-		color: var(--color-link-dark);
-	}
-
-	.select-dropdown__item:hover {
-		background: var(--color-link-5);
-		color: var(--color-neutral-90);
-	}
-}
-
-.product-listing__volume-price-selector {
-	min-width: 100%;
-
-	@include break-medium {
-		min-width: 430px;
-	}
-}
-
-.product-listing__actions-search-and-filter {
-	display: flex;
-	gap: 8px;
-}
-
-.product-listing__empty-result-message {
-	text-align: center;
-	padding: 16px;
-	border-radius: 4px;
-	max-width: 800px;
-	margin: 64px auto 0;
-}
-
-.product-listing__empty-result-message-heading {
-	font-size: 1rem;
-	font-weight: 600;
-}
-
-.product-listing__empty-result-message-description {
-	font-size: rem(14px);
-	color: var(--color-neutral-50);
-}
