--- client/a8c-for-agencies/sections/marketplace/products-overview/style.scss
@@ -1,87 +1,0 @@
-@import "@wordpress/base-styles/breakpoints";
-@import "@wordpress/base-styles/mixins";
-@import "../mixins";
-
-.products-overview__select-license {
-	// :active introduces a 1px border but :focus introduces a 1.5px border which causes the button to move around.
-	border-width: 1.5px !important;
-	border-color: transparent;
-	text-wrap: nowrap;
-
-	@include breakpoint-deprecated("<660px") {
-		width: auto;
-		flex-grow: 1;
-	}
-}
-
-.products-overview .assign-license-step-progress {
-	margin-block-start: 16px;
-	margin-block-end: 32px;
-
-	.step-current .assign-license-step-progress__step-circle {
-		background-color: var(--color-primary-50);
-		border: 2px solid var(--color-primary-50);
-		color: var(--color-text-inverted);
-	}
-}
-
-.products-overview.without-bundle {
-	.hosting-dashboard-layout__header-subtitle {
-		padding-block-end: 16px;
-	}
-}
-
-.a4a-marketplace__header-actions {
-	display: flex;
-	align-items: center;
-	gap: 8px;
-	justify-content: flex-end;
-	& > * {
-		flex-grow: 0;
-	}
-
-	@include breakpoint-deprecated( "<660px" ) {
-		.current-section button {
-			border: none;
-		}
-		justify-content: space-between;
-	}
-}
-
-.products-overview__controls {
-	@include licensing-portal-bottom-action-bar;
-	padding: 14px;
-}
-
-
-.products-overview .hosting-dashboard-layout__header {
-
-	@include breakpoint-deprecated( "<1280px" ) {
-		flex-wrap: nowrap;
-		justify-content: space-between;
-
-	}
-}
-
-.products-overview .hosting-dashboard-layout__body {
-	overflow-y: scroll; // We need to show the scrollbar so we avoid UI from shifting when filtering items.
-}
-
-.products-overview {
-	.hosting-dashboard-layout__top-wrapper {
-		z-index: 10;
-	}
-
-	.select-dropdown__item-text {
-		.content svg {
-			position: relative;
-			top: 5px;
-			margin-inline-end: 4px;
-		}
-	}
-
-	.section-nav.hosting-dashboard-layout__navigation .select-dropdown__header .count {
-		top: 8.5px;
-		margin-inline-end: 8px;
-	}
-}
